requirements.txt
pandas
python-binance
numpy
requests
matplotlib
streamlit
mplfinance
watchdog